fix(xml_parsing): parsing xml added and removed unnecessary files
